Optimized: JSON-RPC types are now defined globally
This was necessary as a step towards enabling a Renderhive message protocol, that uses the Hedera Consensus Service as a transport layer. Those messages include JSON-RPC commands, which – after filtering and verification – are relayed to the node's internal JSON-RPC.
